test(client-s3): move selectObjectContent test from e2e to integ
