+++ clients/client-s3/test/e2e/S3.e2e.spec.ts
@@ -231,46 +231,6 @@ describe("@aws-sdk/client-s3", () => {
     });
   });
 
-  describe("selectObjectContent", () => {
-    const csvFile = `user_name,age
-jsrocks,13
-node4life,22
-esfuture,29`;
-    beforeAll(async () => {
-      Key = `${Date.now()}`;
-      await client.putObject({ Bucket, Key, Body: csvFile });
-    });
-    afterAll(async () => {
-      await client.deleteObject({ Bucket, Key });
-    });
-    it("should succeed", async () => {
-      const { Payload } = await client.selectObjectContent({
-        Bucket,
-        Key,
-        ExpressionType: "SQL",
-        Expression: "SELECT user_name FROM S3Object WHERE cast(age as int) > 20",
-        InputSerialization: {
-          CSV: {
-            FileHeaderInfo: "USE",
-            RecordDelimiter: "\n",
-            FieldDelimiter: ",",
-          },
-        },
-        OutputSerialization: {
-          CSV: {},
-        },
-      });
-      const events: SelectObjectContentEventStream[] = [];
-      for await (const event of Payload!) {
-        events.push(event);
-      }
-      expect(events.length).toEqual(3);
-      expect(new TextDecoder().decode(events[0].Records?.Payload)).toEqual("node4life\nesfuture\n");
-      expect(events[1].Stats?.Details).toBeDefined();
-      expect(events[2].End).toBeDefined();
-    });
-  });
-
   describe("Multi-region access point", () => {
     beforeAll(async () => {
       Key = `${Date.now()}`;
